pyproject.toml
[tool.poetry]
name = "m2ds-data-stream-kafka"
version = "0.1.0"
description = ""
authors = ["Baptiste Pasquier <pasquierbaptiste@hotmail.fr>"]
readme = "README.md"
packages = [{include = "m2ds_data_stream_kafka", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
tweepy = "^4.12.1"
kafka-python = "^2.0.2"
nltk = "^3.7"
Unidecode = "^1.3.6"
colorama = "^0.4.6"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = {extras = ["jupyter"], version = "^22.10.0"}
ipykernel = "^6.17.1"
ipywidgets = ">=7.0,<8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
